sdk/formrecognizer/azure-ai-formrecognizer/samples/v3.2_and_later/requirements.txt
python-dotenv
azure-ai-formrecognizer
